mostly completed
Forgot to be commiting my work. Have been following and learning how to use ai tools and packages to track stock prices of a specific company and print its outlook. It is done for the most part and works. Learned how to create this from neuralnine youtube channel and have been working on this for the last week.